Move CanonicalIdentifier to the identifiers module
